Moved utils
